backend: refactor ctrlblock and rob
* Move RedirectGenerator, Snapshot, DebugMdpInfo etc. in seperate files
