+++ src/main/scala/xiangshan/backend/CtrlBlock.scala
@@ -27,10 +27,12 @@ import xiangshan.backend.decode.{DecodeStage, FusionDecoder, ImmUnion}
 import xiangshan.backend.dispatch.{Dispatch, Dispatch2Rs, DispatchQueue}
 import xiangshan.backend.fu.PFEvent
 import xiangshan.backend.rename.{Rename, RenameTableWrapper}
-import xiangshan.backend.rob.{DebugLSIO, LsTopdownInfo, Rob, RobCSRIO, RobLsqIO, RobPtr}
+import xiangshan.backend.rob.{Rob, RobCSRIO, RobLsqIO, RobPtr}
+import xiangshan.backend.ctrlblock.{DebugLSIO, LsTopdownInfo}
 import xiangshan.frontend.{FtqPtr, FtqRead, Ftq_RF_Components}
 import xiangshan.mem.mdp.{LFST, SSIT, WaitTable}
 import xiangshan.ExceptionNO._
+import xiangshan.backend.ctrlblock._
 import xiangshan.backend.exu.ExuConfig
 import xiangshan.mem.{LsqEnqCtrl, LsqEnqIO}
 
@@ -38,194 +40,6 @@ class CtrlToFtqIO(implicit p: Parameters) extends XSBundle {
   def numRedirect = exuParameters.JmpCnt + exuParameters.AluCnt
   val rob_commits = Vec(CommitWidth, Valid(new RobCommitInfo))
   val redirect = Valid(new Redirect)
-}
-
-class SnapshotPtr(implicit p: Parameters) extends CircularQueuePtr[SnapshotPtr](
-  p => p(XSCoreParamsKey).RenameSnapshotNum
-)
-
-object SnapshotGenerator extends HasCircularQueuePtrHelper {
-  def apply[T <: Data](enqData: T, enq: Bool, deq: Bool, flush: Bool)(implicit p: Parameters): Vec[T] = {
-    val snapshotGen = Module(new SnapshotGenerator(enqData))
-    snapshotGen.io.enq := enq
-    snapshotGen.io.enqData.head := enqData
-    snapshotGen.io.deq := deq
-    snapshotGen.io.flush := flush
-    snapshotGen.io.snapshots
-  }
-}
-
-class SnapshotGenerator[T <: Data](dataType: T)(implicit p: Parameters) extends XSModule
-  with HasCircularQueuePtrHelper {
-
-  class SnapshotGeneratorIO extends Bundle {
-    val enq = Input(Bool())
-    val enqData = Input(Vec(1, chiselTypeOf(dataType))) // make chisel happy
-    val deq = Input(Bool())
-    val flush = Input(Bool())
-    val snapshots = Output(Vec(RenameSnapshotNum, chiselTypeOf(dataType)))
-    val enqPtr = Output(new SnapshotPtr)
-    val deqPtr = Output(new SnapshotPtr)
-    val valids = Output(Vec(RenameSnapshotNum, Bool()))
-  }
-
-  val io = IO(new SnapshotGeneratorIO)
-
-  val snapshots = Reg(Vec(RenameSnapshotNum, chiselTypeOf(dataType)))
-  val snptEnqPtr = RegInit(0.U.asTypeOf(new SnapshotPtr))
-  val snptDeqPtr = RegInit(0.U.asTypeOf(new SnapshotPtr))
-  val snptValids = RegInit(VecInit.fill(RenameSnapshotNum)(false.B))
-
-  io.snapshots := snapshots
-  io.enqPtr := snptEnqPtr
-  io.deqPtr := snptDeqPtr
-  io.valids := snptValids
-
-  when(!isFull(snptEnqPtr, snptDeqPtr) && io.enq) {
-    snapshots(snptEnqPtr.value) := io.enqData.head
-    snptValids(snptEnqPtr.value) := true.B
-    snptEnqPtr := snptEnqPtr + 1.U
-  }
-  when(io.deq) {
-    snptValids(snptDeqPtr.value) := false.B
-    snptDeqPtr := snptDeqPtr + 1.U
-    XSError(isEmpty(snptEnqPtr, snptDeqPtr), "snapshots should not be empty when dequeue!\n")
-  }
-  when(io.flush) {
-    snptValids := 0.U.asTypeOf(snptValids)
-    snptEnqPtr := 0.U.asTypeOf(new SnapshotPtr)
-    snptDeqPtr := 0.U.asTypeOf(new SnapshotPtr)
-  }
-}
-
-class RedirectGenerator(implicit p: Parameters) extends XSModule
-  with HasCircularQueuePtrHelper {
-
-  class RedirectGeneratorIO(implicit p: Parameters) extends XSBundle {
-    def numRedirect = exuParameters.JmpCnt + exuParameters.AluCnt
-    val hartId = Input(UInt(8.W))
-    val exuMispredict = Vec(numRedirect, Flipped(ValidIO(new ExuOutput)))
-    val loadReplay = Flipped(ValidIO(new Redirect))
-    val flush = Input(Bool())
-    val redirectPcRead = new FtqRead(UInt(VAddrBits.W))
-    val stage2Redirect = ValidIO(new Redirect)
-    val stage3Redirect = ValidIO(new Redirect)
-    val memPredUpdate = Output(new MemPredUpdateReq)
-    val memPredPcRead = new FtqRead(UInt(VAddrBits.W)) // read req send form stage 2
-    val isMisspreRedirect = Output(Bool())
-  }
-  val io = IO(new RedirectGeneratorIO)
-  /*
-        LoadQueue  Jump  ALU0  ALU1  ALU2  ALU3   exception    Stage1
-          |         |      |    |     |     |         |
-          |============= reg & compare =====|         |       ========
-                            |                         |
-                            |                         |
-                            |                         |        Stage2
-                            |                         |
-                    redirect (flush backend)          |
-                    |                                 |
-               === reg ===                            |       ========
-                    |                                 |
-                    |----- mux (exception first) -----|        Stage3
-                            |
-                redirect (send to frontend)
-   */
-  def selectOldestRedirect(xs: Seq[Valid[Redirect]]): Vec[Bool] = {
-    val compareVec = (0 until xs.length).map(i => (0 until i).map(j => isAfter(xs(j).bits.robIdx, xs(i).bits.robIdx)))
-    val resultOnehot = VecInit((0 until xs.length).map(i => Cat((0 until xs.length).map(j =>
-      (if (j < i) !xs(j).valid || compareVec(i)(j)
-      else if (j == i) xs(i).valid
-      else !xs(j).valid || !compareVec(j)(i))
-    )).andR))
-    resultOnehot
-  }
-
-  def getRedirect(exuOut: Valid[ExuOutput]): ValidIO[Redirect] = {
-    val redirect = Wire(Valid(new Redirect))
-    redirect.valid := exuOut.valid && exuOut.bits.redirect.cfiUpdate.isMisPred
-    redirect.bits := exuOut.bits.redirect
-    redirect.bits.debugIsCtrl := true.B
-    redirect.bits.debugIsMemVio := false.B
-    redirect
-  }
-
-  val jumpOut = io.exuMispredict.head
-  val allRedirect = VecInit(io.exuMispredict.map(x => getRedirect(x)) :+ io.loadReplay)
-  val oldestOneHot = selectOldestRedirect(allRedirect)
-  val needFlushVec = VecInit(allRedirect.map(_.bits.robIdx.needFlush(io.stage2Redirect) || io.flush))
-  val oldestValid = VecInit(oldestOneHot.zip(needFlushVec).map{ case (v, f) => v && !f }).asUInt.orR
-  val oldestExuOutput = Mux1H(io.exuMispredict.indices.map(oldestOneHot), io.exuMispredict)
-  val oldestRedirect = Mux1H(oldestOneHot, allRedirect)
-  io.isMisspreRedirect := VecInit(io.exuMispredict.map(x => getRedirect(x).valid)).asUInt.orR
-  io.redirectPcRead.ptr := oldestRedirect.bits.ftqIdx
-  io.redirectPcRead.offset := oldestRedirect.bits.ftqOffset
-
-  val s1_jumpTarget = RegEnable(jumpOut.bits.redirect.cfiUpdate.target, jumpOut.valid)
-  val s1_imm12_reg = RegNext(oldestExuOutput.bits.uop.ctrl.imm(11, 0))
-  val s1_pd = RegNext(oldestExuOutput.bits.uop.cf.pd)
-  val s1_redirect_bits_reg = RegNext(oldestRedirect.bits)
-  val s1_redirect_valid_reg = RegNext(oldestValid)
-  val s1_redirect_onehot = RegNext(oldestOneHot)
-
-  // stage1 -> stage2
-  io.stage2Redirect.valid := s1_redirect_valid_reg && !io.flush
-  io.stage2Redirect.bits := s1_redirect_bits_reg
-
-  val s1_isReplay = s1_redirect_onehot.last
-  val s1_isJump = s1_redirect_onehot.head
-  val real_pc = io.redirectPcRead.data
-  val brTarget = real_pc + SignExt(ImmUnion.B.toImm32(s1_imm12_reg), XLEN)
-  val snpc = real_pc + Mux(s1_pd.isRVC, 2.U, 4.U)
-  val target = Mux(s1_isReplay,
-    real_pc, // replay from itself
-    Mux(s1_redirect_bits_reg.cfiUpdate.taken,
-      Mux(s1_isJump, s1_jumpTarget, brTarget),
-      snpc
-    )
-  )
-
-  val stage2CfiUpdate = io.stage2Redirect.bits.cfiUpdate
-  stage2CfiUpdate.pc := real_pc
-  stage2CfiUpdate.pd := s1_pd
-  // stage2CfiUpdate.predTaken := s1_redirect_bits_reg.cfiUpdate.predTaken
-  stage2CfiUpdate.target := target
-  // stage2CfiUpdate.taken := s1_redirect_bits_reg.cfiUpdate.taken
-  // stage2CfiUpdate.isMisPred := s1_redirect_bits_reg.cfiUpdate.isMisPred
-
-  val s2_target = RegEnable(target, s1_redirect_valid_reg)
-  val s2_pc = RegEnable(real_pc, s1_redirect_valid_reg)
-  val s2_redirect_bits_reg = RegEnable(s1_redirect_bits_reg, s1_redirect_valid_reg)
-  val s2_redirect_valid_reg = RegNext(s1_redirect_valid_reg && !io.flush, init = false.B)
-
-  io.stage3Redirect.valid := s2_redirect_valid_reg
-  io.stage3Redirect.bits := s2_redirect_bits_reg
-
-  // get pc from ftq
-  // valid only if redirect is caused by load violation
-  // store_pc is used to update store set
-  val store_pc = io.memPredPcRead(s1_redirect_bits_reg.stFtqIdx, s1_redirect_bits_reg.stFtqOffset)
-
-  // update load violation predictor if load violation redirect triggered
-  io.memPredUpdate.valid := RegNext(s1_isReplay && s1_redirect_valid_reg, init = false.B)
-  // update wait table
-  io.memPredUpdate.waddr := RegNext(XORFold(real_pc(VAddrBits-1, 1), MemPredPCWidth))
-  io.memPredUpdate.wdata := true.B
-  // update store set
-  io.memPredUpdate.ldpc := RegNext(XORFold(real_pc(VAddrBits-1, 1), MemPredPCWidth))
-  // store pc is ready 1 cycle after s1_isReplay is judged
-  io.memPredUpdate.stpc := XORFold(store_pc(VAddrBits-1, 1), MemPredPCWidth)
-
-  // // recover runahead checkpoint if redirect
-  // if (!env.FPGAPlatform) {
-  //   val runahead_redirect = Module(new DifftestRunaheadRedirectEvent)
-  //   runahead_redirect.io.clock := clock
-  //   runahead_redirect.io.coreid := io.hartId
-  //   runahead_redirect.io.valid := io.stage3Redirect.valid
-  //   runahead_redirect.io.pc :=  s2_pc // for debug only
-  //   runahead_redirect.io.target_pc := s2_target // for debug only
-  //   runahead_redirect.io.checkpoint_id := io.stage3Redirect.bits.debug_runahead_checkpoint_id // make sure it is right
-  // }
 }
 
 class CtrlBlock(dpExuConfigs: Seq[Seq[Seq[ExuConfig]]])(implicit p: Parameters) extends LazyModule
