+++ src/main/scala/xiangshan/backend/rename/freelist/BaseFreeList.scala
@@ -20,7 +20,7 @@ import chipsalliance.rocketchip.config.Parameters
 import chisel3._
 import chisel3.util._
 import xiangshan._
-import xiangshan.backend.SnapshotGenerator
+import xiangshan.backend.ctrlblock.SnapshotGenerator
 import utils._
 import utility._
